pyproject.toml
[project]
name = "uiap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=45.0.5",
    "pyserial>=3.5",
    "pyside6>=6.9.1",
    "ymodem",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.uv.sources]
ymodem = { git = "https://github.com/alexwoo1900/ymodem.git" }

[dependency-groups]
build = ["pyinstaller>=6.14.2"]
